requirements.txt
appdirs==1.4.4
asgiref==3.3.4
autopep8==1.5.7
baidu-aip==2.2.18.0
bcrypt==3.2.0
certifi==2019.11.28
cffi==1.14.6
chardet==3.0.4
cryptography==3.4.7
distlib==0.3.1
Django==3.2.8
django-cors-headers==3.8.0
django-cors-middleware==1.5.0
django-rest-framework==0.1.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
filelock==3.0.12
idna==2.8
importlib-metadata==4.0.1
Markdown==3.3.4
numpy==1.21.0
opencv-python==4.5.2.54
panda==0.3.1
pandas==1.3.2
paramiko==2.7.2
Pillow==8.3.0
pipenv==2020.11.15
psycopg2==2.9.1
pycodestyle==2.7.0
pycparser==2.20
PyJWT==1.7.1
PyNaCl==1.4.0
pytesseract==0.3.8
python-dateutil==2.8.2
pytz==2021.1
requests==2.22.0
six==1.15.0
sqlparse==0.4.1
sshtunnel==0.4.0
toml==0.10.2
typing-extensions==3.7.4.3
urllib3==1.25.7
virtualenv==20.4.4
virtualenv-clone==0.5.4
whl==0.0.1
zipp==3.4.1
